# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CryptoSaham Predictor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CryptoAsset {
  id: string;
  symbol: string;
  name: string;
  current_price: number;
  market_cap: number;
  market_cap_rank: number;
  price_change_percentage_24h: number;
  /** @nullable */
  price_change_percentage_7d?: number | null;
  total_volume: number;
  /** @nullable */
  high_24h?: number | null;
  /** @nullable */
  low_24h?: number | null;
  image: string;
  /** @nullable */
  ath?: number | null;
  /** @nullable */
  circulating_supply?: number | null;
}

export interface CryptoDetail {
  id: string;
  symbol: string;
  name: string;
  current_price: number;
  market_cap: number;
  price_change_percentage_24h?: number;
  /** @nullable */
  price_change_percentage_7d?: number | null;
  /** @nullable */
  price_change_percentage_30d?: number | null;
  total_volume?: number;
  /** @nullable */
  high_24h?: number | null;
  /** @nullable */
  low_24h?: number | null;
  image?: string;
  /** @nullable */
  ath?: number | null;
  /** @nullable */
  circulating_supply?: number | null;
  description: string;
  categories: string[];
}

export interface PricePoint {
  timestamp: number;
  price: number;
}

export interface PriceHistory {
  id: string;
  prices: PricePoint[];
}

export interface StockAsset {
  symbol: string;
  name: string;
  price: number;
  change: number;
  changePercent: number;
  /** @nullable */
  volume: number | null;
  /** @nullable */
  marketCap: number | null;
  /** @nullable */
  high?: number | null;
  /** @nullable */
  low?: number | null;
  /** @nullable */
  open?: number | null;
  /** @nullable */
  previousClose?: number | null;
  exchange: string;
  currency?: string;
  /** @nullable */
  logoUrl?: string | null;
}

export interface MarketOverview {
  totalMarketCap: number;
  totalVolume24h: number;
  btcDominance: number;
  ethDominance: number;
  marketCapChange24h: number;
  activeCryptocurrencies: number;
  fearGreedIndex: number;
  fearGreedLabel: string;
}

export interface TrendingCrypto {
  id: string;
  name: string;
  symbol: string;
  rank: number;
  priceChangePercent24h: number;
  image: string;
  /** @nullable */
  currentPrice?: number | null;
}

export interface TrendingStock {
  symbol: string;
  name: string;
  changePercent: number;
  exchange: string;
  /** @nullable */
  price?: number | null;
}

export interface TrendingAssets {
  cryptos: TrendingCrypto[];
  stocks: TrendingStock[];
}

export type NewsArticleType =
  (typeof NewsArticleType)[keyof typeof NewsArticleType];

export const NewsArticleType = {
  crypto: "crypto",
  stock: "stock",
  general: "general",
} as const;

export type NewsArticleSentiment =
  (typeof NewsArticleSentiment)[keyof typeof NewsArticleSentiment];

export const NewsArticleSentiment = {
  positive: "positive",
  negative: "negative",
  neutral: "neutral",
} as const;

export type NewsArticleVotes = {
  positive?: number;
  negative?: number;
  important?: number;
};

export interface NewsArticle {
  id: string;
  title: string;
  url: string;
  publishedAt: string;
  source: string;
  type: NewsArticleType;
  sentiment: NewsArticleSentiment;
  /** @nullable */
  sentimentScore?: number | null;
  relatedAssets?: string[];
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  votes?: NewsArticleVotes;
}

export type PredictionAssetType =
  (typeof PredictionAssetType)[keyof typeof PredictionAssetType];

export const PredictionAssetType = {
  crypto: "crypto",
  stock: "stock",
} as const;

export type PredictionSignal =
  (typeof PredictionSignal)[keyof typeof PredictionSignal];

export const PredictionSignal = {
  strong_buy: "strong_buy",
  buy: "buy",
  neutral: "neutral",
  sell: "sell",
  strong_sell: "strong_sell",
} as const;

export interface Prediction {
  assetId: string;
  assetName: string;
  assetType: PredictionAssetType;
  symbol: string;
  /** @nullable */
  image?: string | null;
  signal: PredictionSignal;
  /** 0-100 confidence score */
  confidence: number;
  /** -1 to 1 sentiment score */
  sentimentScore: number;
  priceChange24h: number;
  /** @nullable */
  priceChange7d?: number | null;
  currentPrice?: number;
  reasons: string[];
  newsCount?: number;
  positiveNews?: number;
  negativeNews?: number;
}

export type PredictionDetailAssetType =
  (typeof PredictionDetailAssetType)[keyof typeof PredictionDetailAssetType];

export const PredictionDetailAssetType = {
  crypto: "crypto",
  stock: "stock",
} as const;

export type PredictionDetailSignal =
  (typeof PredictionDetailSignal)[keyof typeof PredictionDetailSignal];

export const PredictionDetailSignal = {
  strong_buy: "strong_buy",
  buy: "buy",
  neutral: "neutral",
  sell: "sell",
  strong_sell: "strong_sell",
} as const;

export type TechnicalIndicatorsTrend =
  (typeof TechnicalIndicatorsTrend)[keyof typeof TechnicalIndicatorsTrend];

export const TechnicalIndicatorsTrend = {
  bullish: "bullish",
  bearish: "bearish",
  sideways: "sideways",
} as const;

export type TechnicalIndicatorsMomentum =
  (typeof TechnicalIndicatorsMomentum)[keyof typeof TechnicalIndicatorsMomentum];

export const TechnicalIndicatorsMomentum = {
  strong: "strong",
  moderate: "moderate",
  weak: "weak",
} as const;

export type TechnicalIndicatorsVolumeTrend =
  (typeof TechnicalIndicatorsVolumeTrend)[keyof typeof TechnicalIndicatorsVolumeTrend];

export const TechnicalIndicatorsVolumeTrend = {
  increasing: "increasing",
  decreasing: "decreasing",
  stable: "stable",
} as const;

export interface TechnicalIndicators {
  rsi: number;
  trend: TechnicalIndicatorsTrend;
  momentum: TechnicalIndicatorsMomentum;
  volumeTrend: TechnicalIndicatorsVolumeTrend;
  support: number;
  resistance: number;
  /** @nullable */
  movingAverage7d?: number | null;
  /** @nullable */
  movingAverage30d?: number | null;
}

export interface PredictionDetail {
  assetId: string;
  assetName: string;
  assetType: PredictionDetailAssetType;
  symbol: string;
  /** @nullable */
  image?: string | null;
  signal: PredictionDetailSignal;
  confidence: number;
  sentimentScore: number;
  priceChange24h: number;
  /** @nullable */
  priceChange7d?: number | null;
  currentPrice?: number;
  reasons: string[];
  technicalIndicators: TechnicalIndicators;
  newsItems: NewsArticle[];
  newsCount?: number;
  positiveNews?: number;
  negativeNews?: number;
}

export type GetCryptoMarketParams = {
  limit?: number;
  currency?: string;
};

export type GetStockMarketParams = {
  /**
   * Comma-separated stock symbols
   */
  symbols?: string;
};

export type GetNewsParams = {
  type?: GetNewsType;
  limit?: number;
};

export type GetNewsType = (typeof GetNewsType)[keyof typeof GetNewsType];

export const GetNewsType = {
  crypto: "crypto",
  stock: "stock",
  all: "all",
} as const;

export type GetPredictionsParams = {
  type?: GetPredictionsType;
  limit?: number;
};

export type GetPredictionsType =
  (typeof GetPredictionsType)[keyof typeof GetPredictionsType];

export const GetPredictionsType = {
  crypto: "crypto",
  stock: "stock",
  all: "all",
} as const;