# lib/api-zod/src/generated/types/cryptoAsset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CryptoSaham Predictor API
 * OpenAPI spec version: 0.1.0
 */

export interface CryptoAsset {
  id: string;
  symbol: string;
  name: string;
  current_price: number;
  market_cap: number;
  market_cap_rank: number;
  price_change_percentage_24h: number;
  /** @nullable */
  price_change_percentage_7d?: number | null;
  total_volume: number;
  /** @nullable */
  high_24h?: number | null;
  /** @nullable */
  low_24h?: number | null;
  image: string;
  /** @nullable */
  ath?: number | null;
  /** @nullable */
  circulating_supply?: number | null;
}