# lib/api-zod/src/generated/types/marketOverview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CryptoSaham Predictor API
 * OpenAPI spec version: 0.1.0
 */

export interface MarketOverview {
  totalMarketCap: number;
  totalVolume24h: number;
  btcDominance: number;
  ethDominance: number;
  marketCapChange24h: number;
  activeCryptocurrencies: number;
  fearGreedIndex: number;
  fearGreedLabel: string;
}